+++ packages/styles/src/packages/composite/dropdown/rarui-dropdown.types.ts
@@ -1,0 +1,9 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { dropdownStyles } from "./raui-dropdown.css";
+import { StandardLonghandProperties } from "../../../index.types";
+
+export type DropdownVariants = RecipeVariants<typeof dropdownStyles.dropdown>;
+export type DropdownSprinkle = Pick<
+  StandardLonghandProperties,
+  "width" | "maxWidth"
+>;